Fixed warnings in tilestacktool

+++ tilestacktool/mathutils.h
@@ -10,7 +10,7 @@ T limit(T x, T minval, T maxval) {
 }
 
 inline int iround(double x) {
-  return (x > 0.0) ? floor(x + 0.5) : ceil(x - 0.5);
+  return (x > 0.0) ? (int)floor(x + 0.5) : (int)ceil(x - 0.5);
 }
 
 #endif
